requirements.txt
yt-dlp>=2024.0.0
openai-whisper>=20231117
streamlit>=1.28.0
